+++ bgstally/constants.py
@@ -48,3 +48,4 @@ class DiscordPostStyle(str, Enum):
 
 DATETIME_FORMAT_JOURNAL = "%Y-%m-%dT%H:%M:%SZ"
 FOLDER_ASSETS = "assets"
+FOLDER_DATA = "otherdata"