+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 * CMDR Spotting. The plugin now keeps track of the players you target and scan, together with when it happened and in which system. It also looks up any public CMDR and Squadron information on Inara. All this information is presented in a new window where you can review the list of all CMDRs you've targeted. There is also a 'Post to Discord' feature so you can post the CMDR information to your Discord server if you wish (manual only).
 * New format available for Discord posts. The (I think) neater and clearer look uses Discord message embeds. The old text-only format is still available from the settings if you prefer it.
 
+### Changes:
+
+* Additional data files created by BGS-Tally (such as the mission log) are now saved in an `otherdata` subfolder to keep the top level folder as tidy as possible.
+
 ### Bug Fixes:
 
 * BGS-Tally was intentionally omitting secondary INF when a faction was in conflict, but it turns out some mission types can have -ve INF effects on those factions. So we now report all secondary INF.